move classes in lib/epub_preparer.rb to separate files
