+++ lib/epub_sip.rb
@@ -1,0 +1,3 @@
+require 'epub/sip_writer'
+require 'epub/zip_file_writer'
+require 'epub/metadata_extractor'
